chore (error-formatter): moved format methods into a separate file
