+++ src/utils/format.ts
@@ -1,0 +1,33 @@
+import { ValidationError } from '@nestjs/common';
+
+export const mapChildrenToValidationErrors = (
+  error: ValidationError,
+  parentPath?: string,
+): ValidationError[] => {
+  if (!(error.children && error.children.length)) {
+    return [error];
+  }
+  const validationErrors = [];
+  parentPath = parentPath ? `${parentPath}.${error.property}` : error.property;
+  for (const item of error.children) {
+    if (item.children && item.children.length) {
+      validationErrors.push(...mapChildrenToValidationErrors(item, parentPath));
+    }
+    validationErrors.push(prependConstraintsWithParentProp(parentPath, item));
+  }
+  return validationErrors;
+};
+
+const prependConstraintsWithParentProp = (
+  parentPath: string,
+  error: ValidationError,
+): ValidationError => {
+  const constraints = {};
+  for (const key in error.constraints) {
+    constraints[key] = `${parentPath}.${error.constraints[key]}`;
+  }
+  return {
+    ...error,
+    constraints,
+  };
+};
